Fix nil error when using duration {2/3}
Somehow, items become nil when using a duration type. This is the first of 3 fixes I need to pull to make this function. Tested using my version and it worked just fine. This removes the harmless error, which can help with debugging other things you are trying to work on.