Fix command bar being obscured by TextChatService (#259)
There is currently an issue with the new TextChatService where if you're using it instead of the old Lua chat, the chat window will show over the Cmdr bar since it's now living in CoreGui, making it impossible to read. This is the cleanest fix I can think of at the moment; I simply to toggle Enabled on the input bar and chat tray in Window.SetVisible. I have made this fork already in my own game and it seems to work well.

Co-authored-by: eryn L. K <eryn@eryn.io>